requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
av==13.1.0
billiard==4.2.1
build==1.2.2.post1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
distro==1.9.0
Django==5.1.4
django-rest-framework==0.1.0
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
flatbuffers==24.3.25
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
jiter==0.7.1
kombu==5.4.2
livekit==0.18.1
livekit-agents==0.11.3
livekit-api==0.8.0
livekit-plugins-deepgram==0.6.11
livekit-plugins-elevenlabs==0.7.7
livekit-plugins-openai==0.10.7
livekit-plugins-silero==0.7.3
livekit-protocol==0.7.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.6.0
mpmath==1.3.0
multidict==6.1.0
numpy==1.26.4
onnxruntime==1.20.0
openai==1.54.4
packaging==24.2
pillow==10.3.0
pip-tools==7.4.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==5.9.8
psycopg2-binary==2.9.10
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.0
pymongo==4.9.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
redis==5.2.1
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
sympy==1.13.3
tqdm==4.67.0
typer==0.13.1
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
yarl==1.17.1
